pyproject.toml
[project]
name = "weather-bot-docker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.21.0",
    "aiohttp>=3.12.14",
    "python-dotenv>=1.1.1",
]
